.colab/requirements.txt
# Core dependencies
google-generativeai==0.3.2
anthropic==0.7.8
dspy-ai==2.6.27

# PDF processing
PyMuPDF==1.23.8

# Data visualization
plotly==6.3.0
pandas==2.3.1

# Document generation
python-docx==0.8.11
Pillow==10.1.0

# Excel/CSV processing
openpyxl==3.1.2
xlrd==2.0.1

# Additional file processing
chardet==5.2.0

# Geographic data processing (simplified for Colab)
requests==2.32.5

# Jupyter widgets
ipywidgets==8.1.1

# Development dependencies (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
